postgres_to_es/pyproject.toml
[tool.poetry]
name = "postgres_to_es"
version = "0.1.0"
description = ""
authors = ["Alexander Kamyanskiy <alexander.kamyanskiy@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.3"
elasticsearch = "^8.1.1"
python-dotenv = "^0.20.0"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
flake8 = "^4.0.1"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
